rt4539: support 16 bits initial brightness
Use 16 bits rather than 8 bits initial_brightness
to cover entire range for 12 bit brightness control.

Bug: 215412247
Test: light on display
(Cherry picked from commit b5b836a078495d4208fd771a777a77536800e525)

+++ display/backlight/rt4539.h
@@ -36,7 +36,7 @@ struct rt4539_platform_data {
 	u8 current_max;
 	u8 brightness_control;
 	u8 enabled_leds;
-	u8 initial_brightness;
+	u16 initial_brightness;
 	u8 boost_ovp_selection;
 	bool led_short_protection;
 	bool exponential_mapping;
